omstrukturer navansatt-modul
