+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/navansatt/api/NavAnsattRoutes.kt
@@ -3,8 +3,8 @@ package no.nav.mulighetsrommet.api.navansatt.api
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import io.ktor.server.util.*
-import no.nav.mulighetsrommet.api.navansatt.NavAnsattService
 import no.nav.mulighetsrommet.api.navansatt.model.Rolle
+import no.nav.mulighetsrommet.api.navansatt.service.NavAnsattService
 import no.nav.mulighetsrommet.api.plugins.getNavAnsattAzureId
 import no.nav.mulighetsrommet.ktor.extensions.getAccessToken
 import no.nav.mulighetsrommet.tokenprovider.AccessType
