+++ mulighetsrommet-api/src/test/kotlin/no/nav/mulighetsrommet/api/navansatt/service/NavAnsattSyncServiceTest.kt
@@ -1,4 +1,4 @@
-package no.nav.mulighetsrommet.api.navansatt
+package no.nav.mulighetsrommet.api.navansatt.service
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldBeEmpty
